create python module for csv to html
created a python module to read in a csv file saved as "WN2016.csv" and
store it in a list of lists where each list is a row in the table. Then
another python module will create the necessary html tags to create a
webpage. Finally another python module will write the table to HTML
format for any size table. The output will be saved as
"tablewebpage.html"