The first one was wrong
disregard the first one. all of the information should be in home page. the other pages are just part of the template and will be used later. So for it reads the database and generates the building drop down list then a calendar appears then you select a date and then depending on what day of the week it is the function checlBounds () which need to be renamed